requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
starlette==0.37.2
python-dotenv==1.0.1
fpdf2==2.7.9
qrcode[pil]==7.4.2
Pillow==10.3.0
